Проект 22 спринта

+++ commerce/delivery/src/main/java/ru/yandex/practicum/controller/DeliveryController.java
@@ -8,6 +8,7 @@ import ru.yandex.practicum.model.DeliveryDto;
 import ru.yandex.practicum.model.OrderDto;
 import ru.yandex.practicum.service.DeliveryService;
 
+import java.math.BigDecimal;
 import java.util.UUID;
 
 @RestController
@@ -38,7 +39,7 @@ public class DeliveryController implements DeliveryClient {
     }
 
     @Override
-    public Double deliveryCost(OrderDto orderDto) {
+    public BigDecimal deliveryCost(OrderDto orderDto) {
         return deliveryService.deliveryCost(orderDto);
     }
 }
