+++ commerce/order/src/main/java/ru/yandex/practicum/model/Order.java
@@ -6,6 +6,7 @@ import lombok.Builder;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 
+import java.math.BigDecimal;
 import java.util.Map;
 import java.util.UUID;
 
@@ -43,9 +44,9 @@ public class Order {
     @Column(name = "fragile")
     private boolean fragile;
     @Column(name = "total_price")
-    private Double totalPrice;
+    private BigDecimal totalPrice;
     @Column(name = "delivery_price")
-    private Double deliveryPrice;
+    private BigDecimal deliveryPrice;
     @Column(name = "product_price")
-    private Double productPrice;
+    private BigDecimal productPrice;
 }
