+++ commerce/payment/src/main/java/ru/yandex/practicum/service/PaymentService.java
@@ -3,12 +3,13 @@ package ru.yandex.practicum.service;
 import ru.yandex.practicum.model.OrderDto;
 import ru.yandex.practicum.model.PaymentDto;
 
+import java.math.BigDecimal;
 import java.util.UUID;
 
 public interface PaymentService {
     PaymentDto addPayment(OrderDto orderDto);
 
-    Double getTotalCost(OrderDto orderDto);
+    BigDecimal getTotalCost(OrderDto orderDto);
 
     void paymentSuccess(UUID orderId);
